Makefile: enable dynamic memory sharing and RPMB-backed storage for optee os
To measure payloads (linux kernel, dtb, etc...)  and store
measurements into fTPM, fTPM needs to be operational before Linux
kernel is loaded, as such, u-boot needs to provide adequate
secure storage[1] support for optee os from the normal world and
optee os needs to be able to dynamically allocate memory as uboot's
fTPM driver is hardcoded to use dynamic memory sharing as of now.

In our case, the secure storage is backed by RPMB[2], a partition
of the eMMC device with authentication and encrytion support via
a one-time programmable security key. For now a testkey is used
(CFG_RPMB_TESTKEY=y）to make debug easier but to make actual
secure system it's better to derive it from platform specific
secrets and avoid storing the key in memory.

Warning: running a build of optee os with CFG_RPMB_WRITE_KEY=y will
         burn the efuse inside your emmc controller, this is not
         reversible.

Dynamic memory sharing requires the range of memory to be
decleared in the source code of optee and is not enabled by
default at the time of writing, this will be addressed by a later
commit.

[1]: https://optee.readthedocs.io/en/latest/architecture/secure_storage.html#rpmb-secure-storage
[2]: https://en.wikipedia.org/wiki/Replay_Protected_Memory_Block